+++ ICD.Connect.Routing.CrestronPro/ControlSystem/Controls/Microphone/Dmps3MicrophoneDeviceControl.cs
@@ -94,6 +94,9 @@ namespace ICD.Connect.Routing.CrestronPro.ControlSystem.Controls.Microphone
 		{
 			base.Subscribe(parent);
 
+			if (parent == null)
+				return;
+
 #if SIMPLSHARP
 			parent.ControlSystem.MicrophoneChange += ControlSystemOnMicrophoneChange;
 #endif
@@ -106,6 +109,9 @@ namespace ICD.Connect.Routing.CrestronPro.ControlSystem.Controls.Microphone
 		protected override void Unsubscribe(ControlSystemDevice parent)
 		{
 			base.Unsubscribe(parent);
+
+			if (parent == null)
+				return;
 
 #if SIMPLSHARP
 			parent.ControlSystem.MicrophoneChange -= ControlSystemOnMicrophoneChange;